delete dataset on predictor delete

+++ mindsdb/interfaces/model/model_controller.py
@@ -1,7 +1,7 @@
 import os
 import sys
 import time
-from pandas.core.frame import DataFrame
+import json
 import psutil
 import datetime
 from copy import deepcopy
@@ -14,6 +14,7 @@ from lightwood.api.types import ProblemDefinition
 from lightwood import __version__ as lightwood_version
 import numpy as np
 import pandas as pd
+from pandas.core.frame import DataFrame
 import mindsdb_datasources
 
 from mindsdb import __version__ as mindsdb_version
@@ -24,6 +25,7 @@ from mindsdb.utilities.config import Config
 from mindsdb.interfaces.storage.fs import FsStore
 from mindsdb.utilities.log import log
 from mindsdb.interfaces.model.learn_process import LearnProcess, GenerateProcess, FitProcess, UpdateProcess
+from mindsdb.interfaces.datastore.datastore import DataStore
 
 IS_PY36 = sys.version_info[1] <= 6
 
@@ -359,7 +361,19 @@ class ModelController():
         name = f'{company_id}@@@@@{name}'
 
         db_p = db.session.query(db.Predictor).filter_by(company_id=company_id, name=original_name).first()
+        if db_p is None:
+            raise Exception(f"Predictor '{name}' does not exist")
         db.session.delete(db_p)
+        if db_p.datasource_id is not None:
+            try:
+                dataset_record = db.Datasource.query.get(db_p.datasource_id)
+                if (
+                    isinstance(dataset_record.data, str)
+                    and json.loads(dataset_record.data).get('source_type') != 'file'
+                ):
+                    DataStore().delete_datasource(dataset_record.name, company_id)
+            except Exception:
+                pass
         db.session.commit()
 
         DatabaseWrapper(company_id).unregister_predictor(name)
